+++ integration/tests/lib/msg_state.go
@@ -1,4 +1,4 @@
-package common
+package lib
 
 import (
 	"fmt"
