+++ integration/tests/lib/result_approval_state.go
@@ -1,4 +1,4 @@
-package common
+package lib
 
 import (
 	"bytes"
