+++ integration/tests/lib/util.go
@@ -1,4 +1,4 @@
-package common
+package lib
 
 import (
 	"context"
